+++ src/vscp_bootloader.c
@@ -1,6 +1,6 @@
 /* The MIT License (MIT)
  *
- * Copyright (c) 2014 - 2024 Andreas Merkle
+ * Copyright (c) 2014 - 2025 Andreas Merkle
  * http://www.blue-andi.de
  * vscp@blue-andi.de
  *
@@ -88,8 +88,11 @@ typedef enum
 {
     VSCP_BOOTLOADER_MEM_TYPE_PROGRAM = 0,   /**< Program memory */
     VSCP_BOOTLOADER_MEM_TYPE_EEPROM,        /**< Data EEPROM memory */
-    VSCP_BOOTLOADER_MEM_TYPE_CONFIG,        /**< Configuation memory (fuses, etc.) */
-    VSCP_BOOTLOADER_MEM_TYPE_DATA           /**< Data memory (RAM) */
+    VSCP_BOOTLOADER_MEM_TYPE_CONFIG,        /**< Configuation memory (CPU configuration) */
+    VSCP_BOOTLOADER_MEM_TYPE_DATA,          /**< Data memory (RAM) */
+    VSCP_BOOTLOADER_MEM_TYPE_USERID_GUID,   /**< User id / GUID */
+    VSCP_BOOTLOADER_MEM_TYPE_FUSES,         /**< Fuses */
+    VSCP_BOOTLOADER_MEM_TYPE_BOOTLOADER     /**< Bootloader */
 
 } VSCP_BOOTLOADER_MEM_TYPE;
 
@@ -110,17 +113,19 @@ typedef struct
 
 static void vscp_bootloader_simApp(void);
 static void vscp_bootloader_sendNewNodeOnlineEvent(void);
-static void vscp_bootloader_sendAckEnterBootLoader(uint32_t blockSize, uint32_t numBlocks);
-static void vscp_bootloader_sendNakEnterBootLoader(uint8_t errorCode);
+static void vscp_bootloader_sendAckBootLoader(uint32_t blockSize, uint32_t numBlocks);
+static void vscp_bootloader_sendNakBootLoader(uint8_t errorCode);
 static void vscp_bootloader_programmingProcedure(void);
-static void vscp_bootloader_handleProtocolStartBlockDataTransfer(vscp_RxMessage const * const rxMsg, vscp_bootloader_ProgParam * const progParam);
+static void vscp_bootloader_handleProtocolStartBlock(vscp_RxMessage const * const rxMsg, vscp_bootloader_ProgParam * const progParam);
 static void vscp_bootloader_handleProtocolBlockData(vscp_RxMessage const * const rxMsg, vscp_bootloader_ProgParam * const progParam);
 static void vscp_bootloader_handleProtocolProgramDataBlock(vscp_RxMessage const * const rxMsg, vscp_bootloader_ProgParam * const progParam);
 static BOOL vscp_bootloader_handleProtocolDropNicknameId(vscp_RxMessage const * const rxMsg);
-static void vscp_bootloader_sendAckStartBlockDataTransfer(void);
-static void vscp_bootloader_sendNakStartBlockDataTransfer(void);
+static void vscp_bootloader_handleProtocolBootLoaderCheck(vscp_RxMessage const * const rxMsg);
+static void vscp_bootloader_sendAckStartBlock(void);
+static void vscp_bootloader_sendNakStartBlock(void);
 static void vscp_bootloader_sendAckBlockData(uint16_t crc, uint32_t writePtr);
-static void vscp_bootloader_sendNakBlockData(uint8_t errorCode, uint32_t writePtr);
+static void vscp_bootloader_sendAckBlockChunk(void);
+static void vscp_bootloader_sendNakBlockChunk(void);
 static void vscp_bootloader_sendAckProgramBlockData(uint32_t blockNumber);
 static void vscp_bootloader_sendNakProgramBlockData(uint8_t errorCode, uint32_t blockNumber);
 static void vscp_bootloader_sendAckActivateNewImage(void);
@@ -152,8 +157,6 @@ extern void vscp_bootloader_init(void)
 {
     vscp_bl_adapter_init();
     vscp_tp_adapter_init();
-
-    return;
 }
 
 /**
@@ -210,7 +213,7 @@ extern void vscp_bootloader_run(void)
     }
 
     /* Acknowledge the "enter bootloader event", which was received by the application. */
-    vscp_bootloader_sendAckEnterBootLoader(VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE, VSCP_PLATFORM_PROG_MEM_NUM_BLOCKS);
+    vscp_bootloader_sendAckBootLoader(VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE, VSCP_PLATFORM_PROG_MEM_NUM_BLOCKS);
 
     /* Programming procedure */
     vscp_bootloader_programmingProcedure();
@@ -220,8 +223,6 @@ extern void vscp_bootloader_run(void)
 
     vscp_bl_adapter_reboot();
     /* This line never will be reached. */
-
-    return;
 }
 
 /*******************************************************************************
@@ -283,7 +284,7 @@ static void vscp_bootloader_simApp(void)
                         else
                         /* Bootloader programming algorithm not supported */
                         {
-                            vscp_bootloader_sendNakEnterBootLoader(VSCP_BOOTLOADER_ERROR_PROGRAMMING_ALGO_NOT_SUPPORTED);
+                            vscp_bootloader_sendNakBootLoader(VSCP_BOOTLOADER_ERROR_PROGRAMMING_ALGO_NOT_SUPPORTED);
                         }
                     }
                     break;
@@ -296,8 +297,6 @@ static void vscp_bootloader_simApp(void)
         }
     }
     while(FALSE == success);
-
-    return;
 }
 
 /**
@@ -319,12 +318,12 @@ static void vscp_bootloader_sendNewNodeOnlineEvent(void)
 }
 
 /**
- * This function sends a "ACK enter bootloader" event.
+ * This function sends a "ACK bootloader" event.
  *
  * @param[in]   blockSize   Block size
  * @param[in]   numBlocks   Number of blocks
  */
-static void vscp_bootloader_sendAckEnterBootLoader(uint32_t blockSize, uint32_t numBlocks)
+static void vscp_bootloader_sendAckBootLoader(uint32_t blockSize, uint32_t numBlocks)
 {
     vscp_TxMessage  txMsg;
 
@@ -344,16 +343,14 @@ static void vscp_bootloader_sendAckEnterBootLoader(uint32_t blockSize, uint32_t 
     txMsg.data[7]   = (numBlocks >>  0) & 0xff; /* LSB of number of blocks available. */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
- * This function sends a "NACK enter bootloader" event.
+ * This function sends a "NACK bootloader" event.
  *
  * @param[in]  errorCode   User defined error code
  */
-static void vscp_bootloader_sendNakEnterBootLoader(uint8_t errorCode)
+static void vscp_bootloader_sendNakBootLoader(uint8_t errorCode)
 {
     vscp_TxMessage  txMsg;
 
@@ -366,8 +363,6 @@ static void vscp_bootloader_sendNakEnterBootLoader(uint8_t errorCode)
     txMsg.data[0]   = errorCode;    /* User defined error code */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
@@ -390,7 +385,7 @@ static void vscp_bootloader_programmingProcedure(void)
                 switch(rxMsg.vscpType)
                 {
                 case VSCP_TYPE_PROTOCOL_START_BLOCK:
-                    vscp_bootloader_handleProtocolStartBlockDataTransfer(&rxMsg, &progParam);
+                    vscp_bootloader_handleProtocolStartBlock(&rxMsg, &progParam);
                     break;
 
                 case VSCP_TYPE_PROTOCOL_BLOCK_DATA:
@@ -409,6 +404,10 @@ static void vscp_bootloader_programmingProcedure(void)
                     abortFlag = vscp_bootloader_handleProtocolDropNicknameId(&rxMsg);
                     break;
 
+                case VSCP_TYPE_PROTOCOL_BOOT_LOADER_CHECK:
+                    vscp_bootloader_handleProtocolBootLoaderCheck(&rxMsg);
+                    break;
+
                 default:
                     break;
                 }
@@ -416,17 +415,15 @@ static void vscp_bootloader_programmingProcedure(void)
         }
     }
     while(FALSE == abortFlag);
-
-    return;
 }
 
 /**
- * This function handles the start block data transfer event.
+ * This function handles the start block event.
  *
  * @param[in]   rxMsg       Received message
  * @param[in]   progParam   Programming parameter
  */
-static void vscp_bootloader_handleProtocolStartBlockDataTransfer(vscp_RxMessage const * const rxMsg, vscp_bootloader_ProgParam * const progParam)
+static void vscp_bootloader_handleProtocolStartBlock(vscp_RxMessage const * const rxMsg, vscp_bootloader_ProgParam * const progParam)
 {
     if ((NULL == rxMsg) ||
         (NULL == progParam))
@@ -437,7 +434,7 @@ static void vscp_bootloader_handleProtocolStartBlockDataTransfer(vscp_RxMessage 
     if ((4 > rxMsg->dataSize) ||
         (6 < rxMsg->dataSize))
     {
-        vscp_bootloader_sendNakStartBlockDataTransfer();
+        vscp_bootloader_sendNakStartBlock();
     }
     else
     {
@@ -460,27 +457,25 @@ static void vscp_bootloader_handleProtocolStartBlockDataTransfer(vscp_RxMessage 
         }
 
         /* Validate parameters:
-            * - Only program flash is supported.
-            * - Received block number shall be valid.
-            */
+         * - Only program flash is supported.
+         * - Received block number shall be valid.
+         */
         if (VSCP_BOOTLOADER_MEM_TYPE_PROGRAM != memoryType)
         {
-            vscp_bootloader_sendNakStartBlockDataTransfer();
+            vscp_bootloader_sendNakStartBlock();
         }
         else if (VSCP_PLATFORM_PROG_MEM_NUM_BLOCKS <= progParam->blockNumber)
         {
-            vscp_bootloader_sendNakStartBlockDataTransfer();
+            vscp_bootloader_sendNakStartBlock();
         }
         else
         {
-            vscp_bootloader_sendAckStartBlockDataTransfer();
+            vscp_bootloader_sendAckStartBlock();
 
             /* Reset block buffer index */
             progParam->blockBufferIndex = 0;
         }
     }
-
-    return;
 }
 
 /**
@@ -500,14 +495,12 @@ static void vscp_bootloader_handleProtocolBlockData(vscp_RxMessage const * const
     /* Every block data must be a multiple of 8. */
     if (VSCP_L1_DATA_SIZE != rxMsg->dataSize)
     {
-        vscp_bootloader_sendNakBlockData(VSCP_BOOTLOADER_ERROR_INVALID_MESSAGE,
-                                         progParam->blockNumber * VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE);
+        vscp_bootloader_sendNakBlockChunk();
     }
     /* Block buffer already full? */
     else if (VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE < (progParam->blockBufferIndex + rxMsg->dataSize))
     {
-        vscp_bootloader_sendNakBlockData(VSCP_BOOTLOADER_ERROR_INVALID_MESSAGE,
-                                         progParam->blockNumber * VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE);
+        vscp_bootloader_sendNakBlockChunk();
     }
     else
     {
@@ -520,6 +513,8 @@ static void vscp_bootloader_handleProtocolBlockData(vscp_RxMessage const * const
             ++(progParam->blockBufferIndex);
         }
 
+        vscp_bootloader_sendAckBlockChunk();
+
         /* Complete block received? */
         if (VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE <= progParam->blockBufferIndex)
         {
@@ -529,8 +524,6 @@ static void vscp_bootloader_handleProtocolBlockData(vscp_RxMessage const * const
             vscp_bootloader_sendAckBlockData(crcCalculated, progParam->blockNumber * VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE);
         }
     }
-
-    return;
 }
 
 /**
@@ -580,8 +573,6 @@ static void vscp_bootloader_handleProtocolProgramDataBlock(vscp_RxMessage const 
             vscp_bootloader_sendAckProgramBlockData(progParam->blockNumber);
         }
     }
-
-    return;
 }
 
 /**
@@ -613,9 +604,24 @@ static BOOL vscp_bootloader_handleProtocolDropNicknameId(vscp_RxMessage const * 
 }
 
 /**
- * This function sends a "ACK start block data transfer" event.
+ * This function handles the boot loader check event.
+ *
+ * @param[in]   rxMsg   Received message
  */
-static void vscp_bootloader_sendAckStartBlockDataTransfer(void)
+static void vscp_bootloader_handleProtocolBootLoaderCheck(vscp_RxMessage const * const rxMsg)
+{
+    if (NULL == rxMsg)
+    {
+        return;
+    }
+
+    vscp_bootloader_sendAckBootLoader(VSCP_PLATFORM_PROG_MEM_BLOCK_SIZE, VSCP_PLATFORM_PROG_MEM_NUM_BLOCKS);
+}
+
+/**
+ * This function sends a "ACK start block" event.
+ */
+static void vscp_bootloader_sendAckStartBlock(void)
 {
     vscp_TxMessage  txMsg;
 
@@ -627,14 +633,12 @@ static void vscp_bootloader_sendAckStartBlockDataTransfer(void)
     txMsg.dataSize  = 0;    /* 0 byte data */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
- * This function sends a "NACK start block data transfer" event.
+ * This function sends a "NACK start block" event.
  */
-static void vscp_bootloader_sendNakStartBlockDataTransfer(void)
+static void vscp_bootloader_sendNakStartBlock(void)
 {
     vscp_TxMessage  txMsg;
 
@@ -646,8 +650,6 @@ static void vscp_bootloader_sendNakStartBlockDataTransfer(void)
     txMsg.dataSize  = 0;    /* 0 byte data */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
@@ -674,35 +676,40 @@ static void vscp_bootloader_sendAckBlockData(uint16_t crc, uint32_t writePtr)
     txMsg.data[5]   = (uint8_t)((writePtr >>  0u) & 0xFFu);
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
- * This function sends a "NACK block data" event.
- *
- * @param[in]   errorCode   User defined error code
- * @param[in]   writePtr    Write pointer after the last data has been written
+ * This function sends a "ACK block chunk" event.
  */
-static void vscp_bootloader_sendNakBlockData(uint8_t errorCode, uint32_t writePtr)
+static void vscp_bootloader_sendAckBlockChunk(void)
 {
     vscp_TxMessage  txMsg;
 
     txMsg.vscpClass = VSCP_CLASS_L1_PROTOCOL;
-    txMsg.vscpType  = VSCP_TYPE_PROTOCOL_BLOCK_DATA_NACK;
+    txMsg.vscpType  = VSCP_TYPE_PROTOCOL_BLOCK_CHUNK_ACK;
     txMsg.priority  = VSCP_PRIORITY_7_LOW;
     txMsg.oAddr     = vscp_bootloader_nickname;
     txMsg.hardCoded = FALSE;
-    txMsg.dataSize  = 5;    /* 5 byte data */
-    txMsg.data[0]   = errorCode;
-    txMsg.data[1]   = (uint8_t)((writePtr >> 24u) & 0xFFu);
-    txMsg.data[2]   = (uint8_t)((writePtr >> 16u) & 0xFFu);
-    txMsg.data[3]   = (uint8_t)((writePtr >>  8u) & 0xFFu);
-    txMsg.data[4]   = (uint8_t)((writePtr >>  0u) & 0xFFu);
+    txMsg.dataSize  = 0;    /* 0 byte data */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
+}
 
-    return;
+/**
+ * This function sends a "NACK block chunk" event.
+ */
+static void vscp_bootloader_sendNakBlockChunk(void)
+{
+    vscp_TxMessage  txMsg;
+
+    txMsg.vscpClass = VSCP_CLASS_L1_PROTOCOL;
+    txMsg.vscpType  = VSCP_TYPE_PROTOCOL_BLOCK_CHUNK_NACK;
+    txMsg.priority  = VSCP_PRIORITY_7_LOW;
+    txMsg.oAddr     = vscp_bootloader_nickname;
+    txMsg.hardCoded = FALSE;
+    txMsg.dataSize  = 0;    /* 0 byte data */
+
+    (void)vscp_tp_adapter_writeMessage(&txMsg);
 }
 
 /**
@@ -726,8 +733,6 @@ static void vscp_bootloader_sendAckProgramBlockData(uint32_t blockNumber)
     txMsg.data[3]   = (uint8_t)((blockNumber >>  0u) & 0xFFu);
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
@@ -753,8 +758,6 @@ static void vscp_bootloader_sendNakProgramBlockData(uint8_t errorCode, uint32_t 
     txMsg.data[4]   = (uint8_t)((blockNumber >>  0u) & 0xFFu);
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
@@ -772,8 +775,6 @@ static void vscp_bootloader_sendAckActivateNewImage(void)
     txMsg.dataSize  = 0;    /* 0 byte data */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**
@@ -791,8 +792,6 @@ static void vscp_bootloader_sendNakActivateNewImage(void)
     txMsg.dataSize  = 0;    /* 0 byte data */
 
     (void)vscp_tp_adapter_writeMessage(&txMsg);
-
-    return;
 }
 
 /**